+++ src/app/components/PersonalAvatar/PersonalAvatar.stories.tsx
@@ -1,10 +1,18 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import PersonalAvatar from './PersonalAvatar';
+import { MockAuthProvider } from '@/__mocks__/mockAuthProvider';
 
 const meta: Meta<typeof PersonalAvatar> = {
   title: 'Components/PersonalAvatar',
   component: PersonalAvatar,
   tags: ['autodocs'],
+  decorators: [
+    (Story) => (
+      <MockAuthProvider>
+        <Story />
+      </MockAuthProvider>
+    ),
+  ],
 };
 export default meta;